requirements.txt
streamlit
google-cloud-storage
openai
dotenv
google-genia
json5
